requirements.txt
numpy==1.21.0
pandas==1.3.0
torch==1.9.0
torchvision==0.10.0
scikit-learn==0.24.2
matplotlib==3.4.3
Pillow==8.3.1
tqdm==4.62.0
